Update Cargo.lock and move size_and_align_of_dst back
